First Run and Broke Task Chooser
Finally ran it for the first time. Found an issue with SQLiteHelper
where it wouldn’t auto increment the value. The log for addTask doesn’t
show the correct ID but it is the correct one that’s added. Gave code
to Add Time (Timer) and Add Time (Manual). Add Time (Timer) crashes the
app though when it tries to get the Task Chooser Activity set up.
Unsure on what’s going on. Need to look into alternatives / fixes.